Extract PracticeYogaSection to a separate file and move up
