+++ website/src/app/(landing-pages)/page.tsx
@@ -44,22 +44,6 @@ function Page(props: { children: ReactNode; className?: string }) {
 }
 
 export default function IndexPage() {
-  const yogaMan = (
-    <>
-      <Image
-        src={manInBlackClothingPracticingYoga}
-        alt="Man in black clothing practicing yoga"
-        className="rounded-3xl h-96 lg:h-full object-cover"
-      />
-      <YogaIcon
-        className="absolute inset-1/2 -translate-x-1/2 -translate-y-1/2 size-2/3"
-        stroke="white"
-        strokeWidth="0.2"
-        fill="none"
-      />
-    </>
-  );
-
   return (
     <Page className="mx-auto max-w-[90rem] overflow-hidden">
       <Hero
@@ -87,6 +71,7 @@ export default function IndexPage() {
         <HeroDecorationFromLogo logo={<YogaIcon />} />
       </Hero>
 
+      <PracticeYogaSection className="my-24" />
       <ExploreMainProductCards />
       <EverythingHTTPSection />
       <TrulyExtendableSection className="mt-24 mx-4 md:mx-6" />
@@ -138,55 +123,6 @@ export default function IndexPage() {
           </ListItemAnchor>
         ))}
       </RunAnywhereSection>
-
-      <section className="flex my-24 max-lg:flex-col gap-6 lg:gap-24 px-4 xl:px-[120px]">
-        <div className="grow">
-          <Heading as="h2" size="md">
-            Practice Yoga while doing server-work
-          </Heading>
-          <p className="text-green-800 mt-4">
-            Yoga ensures optimal configuration out-of-the-box, enhancing performance and developer
-            workflow.
-          </p>
-          <div className="relative lg:hidden mt-6">{yogaMan}</div>
-          <div className="flex my-12 gap-8 max-sm:flex-col">
-            <InfoCard
-              // @ts-expect-error
-              as={NextLink}
-              heading="Error masking"
-              icon={<Image src={errorWarningBadge} alt="" />}
-              className="flex-1 p-0 md:p-0 bg-transparent hive-focus rounded-md"
-              href="/docs/features/error-masking"
-            >
-              Enhance security by masking errors to prevent sensitive data leaks.
-            </InfoCard>
-            <InfoCard
-              // @ts-expect-error
-              as={NextLink}
-              heading="Health checks"
-              icon={<Image src={pulseLineBadge} alt="" />}
-              className="flex-1 p-0 md:p-0 bg-transparent hive-focus rounded-md"
-              href="/docs/features/health-check"
-            >
-              Built-in health checks to ensure server vitality and readiness.
-            </InfoCard>
-            <InfoCard
-              // @ts-expect-error
-              as={NextLink}
-              heading="GraphiQL Integration"
-              icon={<Image src={graphqlBadge} alt="" />}
-              className="flex-1 p-0 md:p-0 bg-transparent hive-focus rounded-md"
-              href="/docs/features/graphiql"
-            >
-              In-browser IDE for seamless writing, validation, and testing of GraphQL operations.
-            </InfoCard>
-          </div>
-          <CallToAction variant="primary" href="/docs">
-            Learn more
-          </CallToAction>
-        </div>
-        <div className="basis-5/12 shrink-0 relative max-lg:hidden">{yogaMan}</div>
-      </section>
 
       <ToolsAndLibrariesCards />
       {/* TODO: add later */}
